Moved hash into extension method, added test
